UBUNTU: SAUCE: ubuntu_ltp_controllers: bump timeout for memcg_stress test
BugLink: https://bugs.launchpad.net/bugs/1946348

The default 35 minutes timeout defined in memcg_stress_test.sh from
ltp repo seems not enough for this test to run on some instances, use
LTP_TIMEOUT_MUL=2 to double it to 70 minutes.

We will need to bump the overall timeout setting for the autotest
framework as well. Otherwise even with timeout doubled to 70 mins, it
will still gets killed because the timeout for sub-tests was set to
60 minutes.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>