UBUNTU: SAUCE: ubuntu_lxc: use source code from the archive
BugLink: https://bugs.launchpad.net/bugs/1960094

Use the source code from the archive for the test, so that we won't
have a version mismatch issue between the git repo and the package in
our archives caused by the package upload delays.

Patch tested on Focal and this "undefined symbol: strlcat" error does
not exist anymore.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>