UBUNTU: SAUCE: ubuntu_boot: fix return value for kernel_taint_test.py
exit takes only integer args in the range 0 - 255
If we return the taint value read from /proc/sys/kernel/tainted,
it might exceed this range and causing unexpected behaviour.

On a system with kernel warning, e.g. Intel node spitfire with 4.4
$ ./kernel_taint_test.py
Kernel taint value is 512
Taint bit value: 9 (kernel issued warning)
$ echo 0
0

Let's just simplify the return logic of report_failures() to return
either 0 or 1.

We still return -1 when the tainted file does not exist, this will be
handled in the followup patch.

Luckily we do run this test along with log check test, so even with
this test reporting a false negative result, we can still catch
warnings with the log check test.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>