UBUNTU: SAUCE: ubuntu_kernel_selftests: allow ftrace to finish on riscv64 kernel
The ftrace test in ubuntu_kernel_selftests cannot finish properly on
riscv64 kernels and causing the test result in "Incomplete" state.

This is because of the autotest framework timeout handling on RISCV64
instance is incorrect (lp:1940971). According to Dimitri's
investigation 1800 seconds timeout is actually reached in about 10
minutes. And thus the ftracetest test got aborted early, leave the
whole test suite in a bad state.

To solve this, simply disable the timeout to give it a chance to
finish. It will take up to about 35 minutes to run on 5.13 kernel.

The potential risk for this change are:
  * If ftrace test hangs, we will need to wait until the 4 hours
    timeout on jenkins to kick in and terminate the jenkins job. But
    I think it's OK at the moment as all the other tests in
    ubuntu_kernel_selftests will take > 3 hours to run, we're not too
    far from this threshold.
  * Test shows with ftracetest running correctly, it will bring the
    time cost of ubuntu_kernel_selftests to 3h55m, very close to our
    4h bar on bare-metal jenkins. It's possible to see test being
    killed in the future. But there is nothing we can do other than
    skipping some tests.

Patch tested with F-5.13 / I-5.13 / J-5.15 RISCV instance, test can
finish properly with this patch.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>