UBUNTU: SAUCE: Remove zfs-related packages after test
BugLink: https://bugs.launchpad.net/bugs/1892124

Stop zfs daemon and remove zfs related packages after test to prevent
manually provisioned system from booting with out-of-tree zfs modules
and failed with the kernel_tainted test in ubuntu_boot like:
    Kernel taint value is 4097
    Taint bit value: 0 (proprietary module was loaded)
    *   Modules with GPL Incompatible Licenses:
         zfs: CDDL
         zunicode: CDDL
         zcommon: CDDL
         znvpair: CDDL
         zavl: CDDL
         icp: CDDL
    Taint bit value: 12 (externally-built ('out-of-tree') module was loaded)
    *   Modules not in-tree:
         zfs
         zunicode
         zzstd
         zlua
         zcommon
         znvpair
         zavl
         icp
         spl

This is done by adding a sub-test called post-test-zfs-cleanup and
catch it in run_once(). Not using cleanup() because it will be
triggered after each sub test. It's fine for test like
ubuntu_zfs_stress, but it will be complicated for test with multiple
sub-tests e.g. ubuntu_zfs_xfs_generic

The downside of this is that if you're trying to run it again on
the same SUT manully without removing the autotest/client/tmp/
directory, the setup task won't be triggered as the test version
didn't change. To workaround this we will need to move the code in
setup() to initialize() like what we did for ubuntu_sysdig_smoke_test,
to trigger the package installation, or, remove the .version file in
autotest/client/tmp/TESTNAME/

Search criteria to locate tests that need this change:
    grep -r zfsutils-linux *

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>