UBUNTU: SAUCE: ubuntu_boot: add exception for modules on DGX
Server HWE team uses a preseed script to install extra modules on DGX
systems, this is mimicking user-cases but these out-of-tree modules
and unsigned modules will taints the kernel:
    mlx_compat: loading out-of-tree module taints kernel.
    mlx_compat: module verification failed: signature and/or required
                key missing - tainting kernel

Thus cause the kernel_tainted test failing like:
    Kernel taint value is 12288
    Taint bit value: 12 (externally-built ('out-of-tree') module was loaded)
    *   Modules not in-tree:
         mlx5_ib
         ib_uverbs
         ib_core
         mlx5_core
         mlxdevm
         auxiliary
         mlxfw
         mlx_compat
    Taint bit value: 13 (unsigned module was loaded)

Add these modules to an exclusion list so that we don't need to hint
this failure for every cycle.

Note that this failure is only existing on Bionic / Focal with DGX
systems. On Impish / Jammy despite these modules were loaded as well
but they are not tainting the kernel. To incorporate this, limit the
exception check on B/F with DGX systems by looking for "DGX" string
in the product name.

Patch tested with DGX node "exotic-skunk" on B/F, passed as expected.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
[PHLin: remove unnecessary taint bit value print]
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>