UBUNTU: SAUCE: ubuntu_qrt_kernel_security: improve test granularity
BugLink: https://bugs.launchpad.net/bugs/1976344

For now we have only 2 cases in this ubuntu_qrt_kernel_security test:
  * setup
  * test-kernel-security-test.py

All the 125 sub-tests are executed in test-kernel-security-test.py,
it's tidious to read the report every cycle for some specific failure
and hinting everything there is a bit risky for this security related
test.

We should probably ask security team to fix these broken tests or fix
these issues in our kernel, but there is always an exception.

Improve this test suite by running sub test case one-by-one with:
    python2 test-kernel-security.py -v TestClass.TestMethod

To retrieve tests from test-kernel-security.py, Use grep here is more
straightforward then trying to load a module with '-' in the filename.
We will have to grep for the class and the method (test case), then
we parse and constuct a dictionary with them. Downside is this will be
highly dependent to the script structure, should be ok as this won't
change too much under python unittest framework.

Time cost is the tradeoff here, it will increase from 1m30s to about
5m45s, but I think it defintely worth it.

Patch tested with Focal OEM-5.14.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>