UBUNTU: SAUCE: ubuntu_qrt_kernel_aslr_collisions: improve test granularity
Improve the test granularity like how we do it for the
ubuntu_qrt_kernel_security test, run sub test cases one-by-one with:
    python2 test-kernel-aslr-collisions.py -v TestClass.TestMethod

Note that the KernelASLRCollisionsTest.test_stack_collision test will
take about 8 minutes to run, set timeout to 15 minutes to guarantee
it can finish in time.

Patch tested with T-3.13 on bare-metal node "gonzo", and it works as
expected.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>