UBUNTU: SAUCE: ubuntu_nvidia_server_driver: nvidia-fs: fail the job correctly
This commit fixes the false positive of a-c-t-entry.sh, which is a
wrapper to make sure a-c-t invoke our target script in the right
directory context. The wrapper itself should fail correspondingly if the
target script fails, otherwise, there will be false positive of a-c-t
job.

Appending set -x for us to have debugging information in the future by
default, so we can reduce our debugging turn-around time. Not necessary
for returning the right status code of a failed job, but will be very
helpful for post-mortem if the job fails.

Signed-off-by: Taihsiang Ho (tai271828) <taihsiang.ho@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>