UBUNTU: SAUCE: ubuntu_ltp_cve: split test out of ubuntu_ltp
BugLink: https://bugs.launchpad.net/bugs/1983666

Some of the CVE tests will check the kernel taint flag to determine if
the test pass or not.

So if some other test failed before the CVE subset, like the
read_all_sys in the fs subset (bug 1973153), it will cause
false-positive failures.

It can also reduce the ubuntu_ltp execution time, avoid test being
killed by the 4hrs jenkins timeout threshold. Another benefit to pull
this out is that we can have more flexibility to select what we want
to test.

The downside is the time cost for one extra deployment for this job.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>