UBUNTU: SAUCE: change some git clone protocol from git to https
When trying to run tests on PS5, the git protocol is not working even
with proxy enabled because of our firewall rule.

Convert the protocol from git to https for selected tests. This
change should not cause any problem to our current SRU as we have
already been using both git and https protocol for git clone command.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Acked-by: Cory Todd <cory.todd@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>