UBUNTU: SAUCE: ubuntu_boot: reconstruct kernel_tatinted check
BugLink: https://bugs.launchpad.net/1999237

We're using a list to store those modules which will taint the kernel.
However we didn't check why the module is tainted but just ignore them
if they're in the loaded module list.

It will cause a false impression that a module is tainting the kernel
with different issues simply because they're loaded. This was found
while investigating Kinetic RPI boot test result (with database added
in the same logic). Within which only the rpivid_hevc module was
unsigned, but every other modules in the list were printed out for the
unsigned taint as well.

Test output:
  Taint bit value: 13 (unsigned module was loaded)
  Exception made in test script: bcm2835_codec
  Exception made in test script: bcm2835_isp
  Exception made in test script: bcm2835_v4l2
  Exception made in test script: bcm2835_mmal_vchiq
  Exception made in test script: rpivid_hevc
  Exception made in test script: snd_bcm2835
  Exception made in test script: vc_sm_cma
  * Unsigned modules found, but they are expected and OK

Acutal module taint flag:
  snd_bcm2835 - C
  rpivid_hevc - CE
  bcm2835_codec -C
  bcm2835_v4l2 - C
  bcm2835_isp - C
  bcm2835_mmal_vchiq - C
  vc_sm_cma C

Reconstruct kernel_tainted check, use a taint flag dictionary to
filter out known issues for different modules. Ignore them only if the
issue is explicitly recorded in the database.

Also remove the process_GPL_incompatible_modules(), I don't think we
really care if a certain module is using GPL or MIT license. We care
if it's tainting the kernel.

Update the DGX / RPI modules based on actual test result.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>