ubuntu_docker_smoke && ubuntu_fan test: Proxy
BugLink: https://bugs.launchpad.net/bugs/2009351

- Append HTTPS_PROXY and NO_PROXY variables to docker http-proxy.conf
These values are required to allow communication with
registry-1.docker.io/v2/ when in the Openstack Cloud

- In the ubuntu fan test, remove the wget attempt for index.docker.io
The URI was invalid and returned a http 404 error, in this case
triggering the `else` and unsetting the http_proxy.

The check itself was not valid, due to Canonicals internal firewall
rules, index-docker and registry-docker are 2 separate acls and each
need to be configured for the network under test.

Signed-off-by: Sean Feole <sean.feole@canonical.com>
Acked-by: Andrei Gherzan <andrei.gherzan@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>