UBUNTU: SAUCE: Use a more flexible way to get SHA1
We are going to skip rsyncing .git directory for qa-regression-testing
repo, by doing so it can save us about 15 minutes.

To cope with this change, we will need to generate SHA1 information on
our server and copy it to SUT.

Since this information is only helpful when debugging a test issue,
there's no need to block the test if this information is not available.

Make this change into all the other git related tests to make them
consistent.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Cory Todd <cory.todd@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>