UBUNTU: SAUCE: ubuntu_kselftests_ftrace: clone ftrace out from ubuntu_kernel_selftests
ftrace test sometimes will hang and make ubuntu_kernel_selftests ended
up in an incomplete state.

Split it out to improve the test granularity and flexibility and thus
make it easier to debug.

The ftrace test in ubuntu_kernel_selftests is not removed in this
commit, it will be removed after everything is in place and we have
test results coming out from this newly added ubuntu_kselftests_ftrace
test.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Andrei Gherzan <andrei.gherzan@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>