UBUNTU: SAUCE: ubuntu_performance_pts: create wrapper function for getting distro info
In Python 3.8 platform.dist() was removed, so for Focal and later series we need to use
python2 or import the distro module.  Create a simple wrapper function that will try
platform.dist() and import distro if needed.  Use options with distro to allow for
same object to be returned on both routes.

Signed-off-by: Ian May <ian.may@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>