UBUNTU: SAUCE: ubuntu_xfstests_*: use loop device for the test
Use loop device for the test, courtesy to Francis who add this code
on AWS cloud first. This patch was ported from there and rebased
on the xfstest build fix (LP: #2008049), and use /dev/loop0 for
finding which free loop device to use.

By doing so we can avoid destroying data in /dev/sdb accidentally.
Also it can reduce the hardware limit of this test, it will no
longer require a scratch drive to run but just 20G free disk space
as suggested in XFSTests doc.

[1] https://github.com/tytso/xfstests-bld/blob/master/Documentation/what-is-xfstests.md

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>