UBUNTU: SAUCE: ubuntu_ltp_controllers: fail cpuset_hotplug on X-aws-4.4 VM directly
BugLink: https://bugs.launchpad.net/bugs/2026722

cpuset_hotplug will hang on X-aws-4.4 VMs (those with CPU <=2). This
issue can be reproduced all the way back to X-aws-1128, with current
and previous LTP fork.

When this happens, it will cause failure to cpuset_memory as well and
eventually leads to an incomplete test report. Let's fail this test
directly on this kernel to get the test going.

This issue didn't emerge until we begin to run ubuntu_ltp_controllers on
Xenial in recent cycle.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Cory Todd <cory.todd@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Acked-by: Sean Feole <sean.feale@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>