UBUNTU: SAUCE: ubuntu_performance_gpudirect_rdma: porting peer memory test
This commit introduce a peer memory test running on specific machines
with loopback infinite bands. The info of those machines enlisted in
hosts.d. The loopback setup is a must for running this test job.
Expanding host.d is able to support more machines.

The job is expected to run on Jammy with linux-nvidia kernel and above.

The original version of this test job is mainly implemented by
dann frazier <dann.frazier@canonical.com> .

Signed-off-by: Taihsiang Ho (tai271828) <taihsiang.ho@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>