UBUNTU: SAUCE: ubuntu_xilinx: Fix snap dependency for Focal
xlnx-config must be installed from xlnx-config 1.0 track, the install of
xlnx-config in setup() was causing the Focal regression test to fail. For Focal
xlnx-config is installed in the _install_deps but Jammy it is installed
in the test setup(). This updates run-test so that both Jammy and Focal
install xlnx-config in the same place.

This also adds support for the KR260 board which is new for Jammy.

Signed-off-by: Portia Stephens <portia.stephens@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>