UBUNTU: SAUCE: ubuntu_zfs_xfs_generic: reset xfstests-dev to working version
BugLink: https://bugs.launchpad.net/bugs/2029500

Ever since commit 3e7529af ("fstests: add AFS support") was introduced
to xfstests-dev, the "0001-xfstests-add-minimal-support-for-zfs.patch"
patch fails to be applied and the tests are not run. Therefore, rely on
the previous version that makes the patch apply cleanly by resetting to
the corresponding tag.

Signed-off-by: Magali Lemes <magali.lemes@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>