UBUNTU: SAUCE: ubuntu_overlayfs_smoke_test: run just ubuntu_overlayfs_smoke_test.sh on release < Focal
BugLink: https://bugs.launchpad.net/bugs/2032779

We have decided to run just the old ubuntu_overlayfs_smoke_test.sh test
on older release (< Focal). As the purpose of the new overlayfs/shift
tests is to have a better understanding of the custom features that we
bring as SAUCE patches and to make sure that newer kernels don't regress.

This would require changes in the overlay-shiftfs-tests repo as well.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>