UBUNTU: SAUCE: ubuntu_ltp_stable: set timeout multiplier to 5 for hugefork02
BugLink: https://bugs.launchpad.net/bugs/2026543

Tests show this hugefork02 test in hugetlb will take about 90 seconds to
run on ARM64 node wright. If the test did not finish successfully, it
will cause hugemmap05, hugemmap05_1, hugemmap10 and hugeshmctl02 to fail
as well.

Set the timeout multiplier to 5, so it will have 5*30 seconds to finish
the test.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>