UBUNTU: SAUCE: ubuntu_kernel_selftests: bump sub-test build to 90 mins for L/M RISCV
BugLink: https://bugs.launchpad.net/bugs/2040482

The net-build will need approximately 2h40m on 6.2 and 6.5 RISCV
instances, probably because of the increased number of tests.

A 3hr threshold should be a reasonable choice. We can either add
another if statement for RISCV, or simply bump the build timeout
to 90 mins and the arch_scale=2 on RISCV will make it 3hr.

Bumping this timeout should not have any negative effect. As I
don't recall that we have any other build task that failed.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>