UBUNTU: SAUCE: ubuntu_stress_smoke: bail out test early if requirement not met
If we want to bail early (even before the testing tool installation)
base on the requirement check in ubuntu_stress_smoke_test_checks.sh,
we need to move the test build process to a new build_source(), and
handle it just like a normal job in run_once().

This is because setup() will be triggered automatically right after
test execution before anything from run_once(). We can't skip it with
anything in run_once().

And it's not ideal to run the check in setup() before building the
test, firstly is because when you try to run the test again on the
same system, setup() won't be triggered unless the test version has
been bumped or the autotest/client/tmp/ubuntu_stress_smoke directory
has been removed. Secondly we might get confused when setup() fails.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>