UBUNTU: SAUCE: ubuntu_stress_smoke_test: fix init setup and cleanup
The test_dir for the cleanup script was not defined, it will cause job
fail with sut-test failure.

We are calling the init and cleanup script in different places (in the
control file and the ubuntu_stress_smoke_test.py), the $PWD will be
different, set it to $HOME to solve this issue.

Reduce the setup-check and setup-init to just 5 minutes, these tasks
should be quick.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>