UBUNTU: SAUCE: ubuntu_stress_smoke_test: adjust BIOS age requirement to 2011
KERNTT-786
I have check through almost all bare-metal hardware in our MAAS (except
node onza and pignite, they refuse to be deployed). And the oldest
hardware can be dated back to 2011 (onibi and pepe).

Modify the BIOS age limit in ubuntu_stress_smoke_test from 2015 to 2011
to cover them all. By doing so we can avoid picking a hardware that
is not going to work with this test.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>