UBUNTU: SAUCE: ubuntu_kselftests_net: bump test build time to 90m*2 for RISCV
KERNTT-354

With the increase of test cases in the net category, recent test on
M-riscv and M-starfive show that they will take up to 2h24m for the
net-build job to run. Which is way more than the current 60m*2
setting. (*2 is the multiplier for RISCV platforms)

Bump the timeout to 90m*2 to give it a chance to finish on RISCV VMs.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>