UBUNTU: SAUCE: ubuntu_kselftests_net: skip netsniff-ng package on FIPS kernels
BugLink: https://bugs.launchpad.net/bugs/2054609

KERNTT-881
The netsniff-ng depends on the ntp package. But after we enabled the
fipsdevppa for installing FIPS compliant userspace tools, the ntp
installation will fail because it will try to run the MD5 sum which
is prohibited by FIPS.

Drop this package when testing FIPS kernels, the tradeoff is the
altnames.sh test will be skipped.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Magali Lemes <magali.lemes@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>