UBUNTU: SAUCE: ubuntu_boot: rename control file
KERNTT-464
Rename control file from control.ubuntu to control. It's the only way
to allow autotest to recognize the test without specifying where the
control file is.

With this change you can run it with something like:
  autotest/client/autotest-local run ubuntu_boot

This is essential for snap testing as we will need to ship tests
along with autotest.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>