UBUNTU: SAUCE: ubuntu_kernel_selftests: disable ptrace/get_set_sud if CONFIG_GENERIC_ENTRY=n
BugLink: https://bugs.launchpad.net/bugs/2044079

The ptrace requests:

  PTRACE_SET_SYSCALL_USER_DISPATCH_CONFIG
  PTRACE_GET_SYSCALL_USER_DISPATCH_CONFIG

... are supported only for architectures with CONFIG_GENERIC_ENTRY=y,
which are x86, s390, riscv, loongarch as of Linux 6.8.

For the other architectures, the ptrace/get_set_sud kselftest always
fails. It is tedious to give a hint to every platform.

Check /boot/config-$(uname), and if CONFIG_GENERIC_ENTRY=y is not found,
let's disable this test.

Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
[PHLin: amend BugLink format and add test name to the title]
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>