UBUNTU: SAUCE: ubuntu_performance_power: fix build issue on Jammy+
BugLink: https://bugs.launchpad.net/bugs/1959090

We need to use a newer version of stress-ng on Jammy+, otherwise the
build will fail with:
  stress-bad-altstack.c:30:32: error: initializer element is not constant
   30 | static const size_t stack_sz = MINSIGSTKSZ;
      | ^~~~~~~~~~~
  make[1]: *** [Makefile:352: stress-bad-altstack.o] Error 1
  make: *** [Makefile:343: all] Error 2

Experiment shows version with tag V0.18.00 works fine on both Noble
and Jammmy (V0.14.04 is the minimum required version on Noble). Use
this version to fix the build issue.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Kevin Becker <kevin.becker@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>