UBUNTU: SAUCE: ubuntu_performance_lkp: fix build issue on Noble
There are several issues when trying to run the LKP test suite on Noble:
1. Some packages like libpython2.7 are missing on Noble
2. When trying to install the iperf3 package on Noble, it will drop into
   an interactive mode asking if iper3 should be started as a daemon.

These issues have been addressed in the upstream test suite. So we just
need to use the correct commit for Noble.

Note that for the special case on node akis we will need to review its
if statement if we are going to test that node with Noble.

V2: fix the lkp_commit variable typo.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Kevin Becker <kevin.becker@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>