UBUNTU: SAUCE: ubuntu_performance_iperf3: add config to support DGXH100
DGXH100's line rate is 400Gb over ethernet on its CX-7 network adapters.
Expect at least 90% of it when we have enough iperf instances; cert
found 15 instances to be the sweet spot.

Enabling LRO via the ethtool utility does not appear to work for the
CX-7 adapters on our DGXH100 instance, so it needs to be skipped.

This makes expected_throughput and whether or not to explicitly set LRO
on configurable.

Signed-off-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>