UBUNTU: SAUCE: ubuntu_stress_smoke_test: fix cgroup2 check
BugLink: https://bugs.launchpad.net/bugs/1950630

When grep for /sys/fs/cgroup from /proc/mounts to determine whether
we should use cgroup v1 or v2 for the test, it will return multiple
lines on B/F and leads to a "[: too many arguments" error.

This issue is not affecting the test, in this case it will use cgroup
v1 for the test on B/F. cgroup v2 memory.max setting does not work on
them as the attempt to creat a /sys/fs/cgroup/stress-ng-test directory
will failed because of permission denied.

This is the line that contains cgroup2 in /proc/mounts:
 * T
     None
 * X
     None
 * B
     cgroup /sys/fs/cgroup/unified cgroup2 rw,nosuid,nodev,noexec,relatime 0 0
 * F
     cgroup2 /sys/fs/cgroup/unified cgroup2 rw,nosuid,nodev,noexec,relatime,nsdelegate 0 0
 * J
     cgroup2 /sys/fs/cgroup cgroup2 rw,nosuid,nodev,noexec,relatime,nsdelegate,memory_recursiveprot 0 0
 * N
     cgroup2 /sys/fs/cgroup cgroup2 rw,nosuid,nodev,noexec,relatime,nsdelegate,memory_recursiveprot 0 0

Fix this detection by looking for "cgroup2 /sys/fs/cgroup " and trim
it down to "cgroup2". So that T/X/B/F will still use v1, and J/N will
still use v2.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>