UBUNTU: SAUCE: rt_tests_cyclictest: Print stats in format used by maas-scrape-influxdb.sh
Calculate minimum, maximum and average latency from the test, using the
final results from cyclictest. These are now printed in a format that
can be imported into InfluxDB. The calculated maximum is now used for
the pass/fail criteria instead of using a break argument on the test.
This allows us to have a failure and retain the full latency info
instead of ending the test early upon a failure.

Signed-off-by: Kevin Becker <kevin.becker@canonical.com>
Acked-by: Francis Ginther <francis.ginther@canonical.com>
Signed-off-by: Kevin Becker <kevin.becker@canonical.com>