UBUNTU: SAUCE: ubuntu_performance_pts: set timeout for apache, crafty to 15 mins
BugLink: https://bugs.launchpad.net/bugs/2029373
Result shows the apache test will need about 12.5 minutes to run and
the crafty test will need about 10 minutes. Which are both higher than
the current 10 minutes setting. Bump the timeout to 15 minutes.

Note that these tests are marked as "NOTRUN: test not run, no data" in
the end. So adjusting this timeout threshold is not lowering the bar
for this performance test.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Kevin Becker <kevin.becker@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>