UBUNTU: SAUCE: ubuntu_futex: use mirror on launchpad
In sru-s20240805, this test failed with git clone command on Trusty:
  fatal: unable to access 'https://git.kernel.org/..../dvhart/futextest.git/':
  gnutls_handshake() failed: Handshake failed

Since the git clone command is still working for repositories hosted on
launchpad, a workaround for this is to create a mirror there. The
maintenance cost should be low as this repositories haven't been updated
for years.

Note that we decided to use https instead of git protocol since commit
2b56e36e UBUNTU: SAUCE: use https protocol for git clone, because of
different firewall restrictions here and there.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Manuel Diewald <manuel.diewald@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>