UBUNTU: SAUCE: ubuntu_lxc: install libdbus-1-dev for Noble+
BugLink: https://bugs.launchpad.net/bugs/2083805

Oracular needs libdbus-1-dev package to build this test. This package is
available for Noble and there is no side-effect with it installed there.
Let's reuse the existing if statement for newer releases.

The if statement for Mantic can be removed as it is not supported anymore.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Sean Feole <sean.feole@canonical.com>