UBUNTU: SAUCE: Remove docker.io after test for fan/docker/overlayfs smoke tests
KERNTT-1167

Certification team has notified us the installed docker.io package is
breaking their tests on manually provisioned MTK instances. The grep
result indicates the following  tests will be installing this package:
  * ubuntu_docker_smoke_test
  * ubuntu_fan_smoke_test
  * ubuntu_k8s_unit_tests

We just run docker and fan smoke tests on the MTK kernel.

Add a cleanup() to remove it.

Apart from these, ubuntu_overlayfs_smoke_test needs be modified too.
The docker.io package will be installed with an install-deps script
in overlay-shiftfs-tests repo. As we won't be able to know which
sub-test is the last test for us to trigger cleanup(), add a new
post-test-cleanup test case for this purpose.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>