UBUNTU: SAUCE: ubuntu_nvidia_server_driver: remove nvidia driver packages in a single transaction
Removing packages one at a time is unstable, and can result in alternative
dependencies being pulled in that are not later removed.

For example, suppose linux-modules-nvidia-550-server-nvidia and
linux-modules-nvidia-550-server-5.15.0-1067-nvidia are installed, and
linux-modules-nvidia-550-server-5.15.0-1068-nvidia is available. When
linux-modules-nvidia-550-server-5.15.0-1067-nvidia is removed,
linux-modules-nvidia-550-server-5.15.0-1068-nvidia is installed to replace it.
This means kernel modules are left installed and uninstall_all_nvidia_mod_pkgs
reports failure.

Instead, remove everything in a single transaction so this does not happen.

Signed-off-by: Jacob Martin <jacob.martin@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>