UBUNTU: SAUCE: ubuntu_ltp_syscalls: stop systemd-timesyncd for clock_settime03
BugLink: https://bugs.launchpad.net/bugs/1939016

The goal of clock_settime03 test is to verify the y2038 vulnerability
(i.e. max value that can be stored in s32). It sets the system time to
just before y2038 time and adds a timer to expire just after crossing it.

This test start failing with timeout on some instances since impish 5.11.
Experiment shows with systemd-timesyncd stopped momentarily for this test
will make it become more reliable.

Patch verified on J/N/O generic kernel with Openstack instances and it is
working as expected.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Sean Feole <sean.feole@canonical.com>