UBUNTU: SAUCE: ubuntu_boot: fix UC environment detection
KERNTT-1314
In commit 51ee86a1f8 "UBUNTU: SAUCE: ubuntu_boot: use /etc/lsb-release for
retriving OS information" we switched from /etc/os-release to /etc/lsb-release
because the former will be affected by the base snap of the Python snap that
we installed for autotest.

However, tests show the content in /etc/lsb-release is not consistent across
different UC releases. Here is the output from UC22:
  DISTRIB_ID=Ubuntu
  DISTRIB_RELEASE=22.04
  DISTRIB_CODENAME=jammy
  DISTRIB_DESCRIPTION="Ubuntu 22.04.5 LTS"

It's not guaranteed to have "Ubuntu Core" string there.

Since we just need to determine if we are running on a Ubuntu Core system to
decide:
  * If the apt commands should be called
  * Locate the kernel config in the right place
I think it's still OK to switch back to /etc/os-release as we not checking
version number with it, but we can tell if it's a UC system.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>