UBUNTU: SAUCE: ubuntu_cve_kernel: CVE-2016-313{4,5}: define _GNU_SOURCE
Starting with Oracular, the CVE-2016-313{4,5} tests fail to build with
"implicit declaration of function ‘unshare’" and
"implicit declaration of function ‘clone’". This is due to GCC 14
treating implicit-function-declaration as an error. Previous releases
have this warning supressed due to the `-w` option being passed to
gcc, so we never notice it.
The synopsis of clone(2) and unshare(2) specify that `#define
_GNU_SOURCE` should be added alongside `#include <sched.h>` in order to
use these functions. Fix the build failures by following that.

Signed-off-by: Magali Lemes <magali.lemes@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>