UBUNTU: SAUCE: ubuntu_overlayfs_smoke_test: bump the setup timeout to 30mins
KERNTT-1356
This test will need to install install lxd and refresh it to a specific
channel, this might take a while.

Test in sru-20250317 shows on o/oracle-64k VM.Standatd.A1.Flex, it will
take about 18 minutes to complete this setup task.

Bump the timeout threshold to 30 minutes to give it more room.

Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Sean Feole <sean.feole@canonical.com>
Signed-off-by: Sean Feole <sean.feole@canonical.com>