UBUNTU: SAUCE: ubuntu_nvidia_fs: major updates to the test
Test reworked to remove dependency on VMs.

Logic to select NVMe drives updated to include multipath NVMe drives

Removed depenency on nvidia-docker2 as it is replaced by
nvidia-container-toolkit

For gds-tools versions > 12.4, gds/samples repo, which is used to compile
and run the tests has been removed. However, it is found in MagnumIO Github
repo. So a check has been added to select the MagnumIO repo if the version of
gds-tools > 12.4

Signed-off-by: Abdur Rahman <abdur.rahman@canonical.com>
Acked-by: Po-Hsu Lin <po-hsu.lin@canonical.com>
Acked-by: Jacob Martin <jacob.martin@canonical.com>
Signed-off-by: Po-Hsu Lin <po-hsu.lin@canonical.com>